% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_demo_2019}
\alias{census_demo_2019}
\title{Key Winnipeg demographics information for 100% data
and total data. Data is at the finest geographical
resolution (Dissemination Area---DA) available. Admittedly not
complete, and based on variables of interest in my work.}
\format{A data frame with 71 columns:
\describe{
\item{geo_uid}{}
\item{population_2016}{}
\item{total_private_dwellings}{}
\item{private_dwellings_occupied_by_usual_residents}{}
\item{land_area_in_square_kilometres}{}
\item{occupied_private_dwellings_by_structural_type_of_dwelling_data}{}
\item{apartment_in_a_building_that_has_five_or_more_storeys}{}
\item{semi_detached_house}{}
\item{row_house}{}
\item{apartment_or_flat_in_a_duplex}{}
\item{apartment_in_a_building_that_has_fewer_than_five_storeys}{}
\item{other_single_attached_house}{}
\item{movable_dwelling}{}
\item{private_households_by_household_size}{}
\item{x1_person}{}
\item{x2_persons}{}
\item{x3_persons}{}
\item{x4_persons}{}
\item{x5_or_more_persons}{}
\item{age_stats}{}
\item{x0_to_4_years}{}
\item{x5_to_9_years}{}
\item{x10_to_14_years}{}
\item{x15_to_19_years}{}
\item{x20_to_24_years}{}
\item{x25_to_29_years}{}
\item{x30_to_34_years}{}
\item{x35_to_39_years}{}
\item{x40_to_44_years}{}
\item{x45_to_49_years}{}
\item{x50_to_54_years}{}
\item{x55_to_59_years}{}
\item{x60_to_64_years}{}
\item{x65_years_and_over}{}
\item{total_marital_status_for_the_population_aged_15_years_and_over_100_percent_data}{}
\item{married_or_living_common_law}{}
\item{not_married_and_not_living_common_law}{}
\item{total_number_of_census_families_in_private_households_100_percent_data}{}
\item{total_couple_census_families_in_private_households_100_percent_data}{}
\item{couples_without_children}{}
\item{couples_with_children}{}
\item{x1_child}{}
\item{x2_children}{}
\item{x3_or_more_children}{}
\item{total_lone_parent_census_families_in_private_households_100_percent_data}{}
\item{x1_child_2}{}
\item{x2_children_2}{}
\item{x3_or_more_children_2}{}
\item{total_persons_not_in_census_families_in_private_households_100_percent_data}{}
\item{official_languages}{}
\item{non_official_languages}{}
\item{total_total_income_groups_in_2015_for_the_population_aged_15_years_and_over_in_private_households_100_percent_data}{}
\item{without_total_income}{}
\item{under_10_000_including_loss}{}
\item{x10_000_to_19_999}{}
\item{x20_000_to_29_999}{}
\item{x30_000_to_39_999}{}
\item{x40_000_to_49_999}{}
\item{x50_000_to_59_999}{}
\item{x60_000_to_69_999}{}
\item{x70_000_to_79_999}{}
\item{x80_000_to_89_999}{}
\item{x90_000_to_99_999}{}
\item{x100_000_to_149_999}{}
\item{x150_000_and_over}{}
\item{total_income_statistics_in_2015_for_private_households_by_household_size_100_percent_data}{}
\item{median_total_income_of_households_in_2015}{}
\item{total_low_income_status_in_2015_for_the_population_in_private_households_to_whom_low_income_concepts_are_applicable_100_percent_data}{}
\item{x0_to_17_years}{}
\item{x18_to_64_years}{}
\item{x65_years_and_over_2}{}
}}
\source{
Statistics Canada via (cancensus)[https://mountainmath.github.io/cancensus/articles/cancensus.html] package.
\url{https://cran.r-project.org/web/packages/cancensus/index.html}
}
\usage{
census_demo_2019
}
\description{
Key Winnipeg demographics information for 100% data
and total data. Data is at the finest geographical
resolution (Dissemination Area---DA) available. Admittedly not
complete, and based on variables of interest in my work.
}
\examples{
\dontrun{
 census_demo_2019
}
}
\keyword{datasets}
