% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime_aug_2019}
\alias{crime_aug_2019}
\title{Winnipeg police crime historical crime statistics up to August 2019
pulled on 2019-12-02. Original csv names standardized with `janitor`'s
`clean_names` function. Descriptions are my best guess.}
\format{A data frame with 9 columns:
\describe{
\item{community}{Community.}
\item{count}{Number of crimes.}
\item{level_3_not_displayed}{Description of crime.}
\item{level_4}{Description of drug offence where applicable.}
\item{neighbourhoods}{Neigbourhood of crime's occurance.}
\item{number_of_records}{Number of records, note can be distinct from count.}
\item{offence_category}{Offence category.}
\item{offence}{Specific offence.}
\item{report_date}{Character string report date in m/d/yyyy format.}
}}
\source{
Winnipeg police crime map for public use
\url{https://www.winnipeg.ca/police/crimestat/viewMap.stm}
}
\usage{
crime_aug_2019
}
\description{
Winnipeg police crime historical crime statistics up to August 2019
pulled on 2019-12-02. Original csv names standardized with `janitor`'s
`clean_names` function. Descriptions are my best guess.
}
\examples{
\dontrun{
 crime_aug_2019
}
}
\keyword{datasets}
