% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_open_data.R
\name{get_open_data}
\alias{get_open_data}
\title{City of Winnipeg Open Data API}
\usage{
get_open_data(type = c("parking", "2018 polling geoms", "air quality",
  "tax parcels", "trees", "building permits", "plow schedule",
  "parking bans", "parks", "park assets", "council expenses",
  "public notices", "mosquito traps", "address coords", "permit details",
  "burials", "dispositions", "311", "lane closures"), app_token = NA,
  email = NA, password = NA, ...)
}
\arguments{
\item{app_token}{Optional app_token parameter for users with a city
of Winnipeg Open Data account available at
(https://data.winnipeg.ca/)}

\item{email}{Optional email parameter for users with a city
of Winnipeg Open Data account available at
(https://data.winnipeg.ca/)}

\item{password}{Optional password parameter for users with a city
of Winnipeg Open Data account available at
(https://data.winnipeg.ca/)}
}
\value{
A dataframe containing the relvant data set.
}
\description{
This function uses the \code{RSocrata} package to get Winnipeg
City data from the Winnipeg Open Data site \code{data.winnipeg.ca}.
Users with a City of Winnipeg open data account can also use their
login credentials to have full (non-throttled) access to the API.
}
\examples{
\dontrun{
    get_open_data("trees")
    get_open_data("parking", app_token = SOME_API_TOKEN,
        email = FAKE_EMAIL,
        password = FAKE_PASSWORD)
}

}
