% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_WPA_Stations.R
\name{get_WPA_stations}
\alias{get_WPA_stations}
\title{Winnipeg WPA Paystations}
\usage{
get_WPA_stations()
}
\value{
A dataframe containing data for all WPA parking
    stations in Winnipeg such as \code{available_space},
    \code{hourly_rate}, \code{timelimit}, location, and
    others.
}
\description{
This function uses the \code{RSocrata} package to get Winnipeg
WPA Paystation data from the Winnipeg Open Data site
\code{data.winnipeg.ca}.
}
\examples{
\dontrun{
    get_WPA_stations()
}

}
