% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mbnc}
\alias{mbnc}
\title{Municipal Benchmarking Network Canada (MBNC) dataset. MBNC collects and
distributes efficiency statistics on participating canadian municipalities.
Data pulled on 2019-12-27.
Original names standardized with the \code{janitor::clean_names}
function.}
\format{A data frame with 21 columns:
\describe{
\item{measure_type_cd}{MBNC classification of measurement.}
\item{municipality_id}{MBNC unique municipality identifier.}
\item{program_area}{MBNC classification of measurement area.}
\item{municipality}{Name of municipality.}
\item{section_desc}{MBNC desciption of of measurement area.}
\item{measure_number}{Unique identifier of measure.}
\item{measure_name}{Name of measure (non-unique).}
\item{numerator}{Numerator of measure (where applicable).}
\item{denominator}{Denominator of measure (where applicable).}
\item{result}{Either the measured result or numerator/denominator (where applicable).}
\item{numerator_definition}{Definition of the numerator.}
\item{denominator_definition}{Definition of the denominator.}
\item{notes}{Further notes on the measure.}
\item{mpmp_measure}{Whether the measure is used in mpmp.}
\item{cao_measure}{Whether the measure is used in cao.}
\item{measure_category}{Broad measure category.}
\item{year_created}{First Year measure was instituted.}
\item{other}{Other notes and peculiarities in the measure.}
\item{influencing_factor}{Things contributing to the measure result.}
\item{measure}{Verbose name of measure with ID. Paste of measure_name and measure_number.}
}}
\source{
Municipal Benchmarking Network Canada. Public measures.
\url{http://mbncanada.ca/}
}
\usage{
mbnc
}
\description{
Municipal Benchmarking Network Canada (MBNC) dataset. MBNC collects and
distributes efficiency statistics on participating canadian municipalities.
Data pulled on 2019-12-27.
Original names standardized with the \code{janitor::clean_names}
function.
}
\examples{
\dontrun{
 data(mbnc)
}
\dontrun{
# To tidy the dataset & remove notes
tidy_mbnc <- long_MBNC \%>\%
select(municipality, fiscal_year, measure, result) \%>\%
pivot_wider(names_from = measure, values_from = result)
}
}
\keyword{datasets}
