% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighbourhoods_2017}
\alias{neighbourhoods_2017}
\title{Boundaries of neighbourhood characterization areas (aka NCA)
as polygons. Neighbourhood characterization areas are used
to delineate neighbourhoods for city planning purposes.}
\format{A data frame with 3 columns:
\describe{
 \item{id}{Neighbourhood id number.}
 \item{name}{Neighbourhood name.}
 \item{geometry}{Neighbourhood shapefile geometry.}
}}
\source{
(Winnipeg Open Data)[https://data.winnipeg.ca/].
\url{https://data.winnipeg.ca/City-Planning/Neighbourhood/fen6-iygi}
}
\usage{
neighbourhoods_2017
}
\description{
Boundaries of neighbourhood characterization areas (aka NCA)
as polygons. Neighbourhood characterization areas are used
to delineate neighbourhoods for city planning purposes.
}
\examples{
\dontrun{
 transport_mode
}
}
\keyword{datasets}
