% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_open_data.R
\name{search_open_data}
\alias{search_open_data}
\title{Search Data Sets Available from City of Winnipeg Open Data API}
\usage{
search_open_data(contains = NA, updated_after = "2015-01-01",
  updated_before = as.character(Sys.Date()),
  issued_after = "2015-01-01",
  issued_before = as.character(Sys.Date()))
}
\arguments{
\item{contains}{Optional expression to grep fields in metadata with.}

\item{updated_after}{Optional date string representing the earliest the
data set being searched has been updated.}

\item{updated_before}{Optional date string representing the most recent the
data set being searched has been updated.}

\item{issued_after}{Optional date string representing the earliest the
data set being search could have been initially issued.}

\item{issued_before}{Optional date string representing the latest the
data set being search could have been initially issued.}
}
\value{
A dataframe containing the open data set(s) matching input parameters.
}
\description{
This function uses the \code{jsonlite} package to get metadata on the
City of Winnipeg Open Data site \code{data.winnipeg.ca}. Returns
a data frame with descriptions of data and code to retrieve data sets
using get_open_data, or through the open data Socrata API.
}
\examples{
\dontrun{
    # Returns all metadata as a data frame
    search_open_data()

    # Examines data sets whose title or description contains 'trees'
    # ignoring case
    search_open_data('trees')

    # Get recently updated data sets
    search_open_data(updated_after = '2019-01-01')

    # Return API short codes for searched data sets
    search_open_data(contains = 'parking')$
}

}
