% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transit.R
\name{stop_flatten}
\alias{stop_flatten}
\title{Transit stop data flattener}
\usage{
stop_flatten(list)
}
\arguments{
\item{list}{is the result of the list that results form
parsing the Transit stops JSON data into R.}
}
\value{
A flattend dataframe containing stops data.
}
\description{
The stop_flatten helper-function flattens JSON from
[Winnipeg Transit API v3](https://api.winnipegtransit.com/home/api/v3)
into a tidy format. For use with rjson::fromJSON.
}
\examples{
\dontrun{
    stop_flatten(
         rjson::fromJSON(<API_call_string>)
         )
}

}
