% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wpg_electoral_wards_2018}
\alias{wpg_electoral_wards_2018}
\title{Winnipeg 2018 Electoral Wards}
\format{A data frame with 6 columns:
\describe{
 \item{councillor}{Self-explanatory.}
 \item{name}{Ward name.}
 \item{number}{Ward number.}
 \item{phone}{Councillor phone number.}
 \item{website}{Councillor website.}
 \item{geometry}{Shapefile geometry.}
}}
\source{
(Winnipeg Open Data)[https://data.winnipeg.ca/].
\url{https://data.winnipeg.ca/Council-Services/Electoral-Ward/ede3-teb8}
}
\usage{
wpg_electoral_wards_2018
}
\description{
Winnipeg 2018 Electoral Wards
}
\examples{
\dontrun{
 wpg_electoral_wards_2018
}
}
\keyword{datasets}
