% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpg_weather.R
\name{wpg_weather}
\alias{wpg_weather}
\title{Winnipeg Weather}
\usage{
wpg_weather(station = c("wpg", "forks", "airport"),
  start = as.character(Sys.Date() - 365),
  end = as.character(Sys.Date()), metric = TRUE)
}
\arguments{
\item{station}{a Winnipeg weather station name in English}

\item{start}{the start date of the weather period}

\item{end}{the end date of the weather period}

\item{metric}{whether the output should be in metric}
}
\value{
hourly Winnipeg weather data set including precipitation
   and temperature
}
\description{
This function uses the \code{reim} package to load Winnipeg
weather abstracting station names. Measurements can be
converted to metric using \code{meathermetrics}
}
\examples{
\dontrun{
    wpg_weather(station = "forks",
                start = '2018-01-01',
                end = '2018-02-28',
                metric = TRUE)
    wpg_weather(station = "airport",
                start = '2018-01-01')
}

}
