% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yrly_crime_2018}
\alias{yrly_crime_2018}
\title{Yearly incident based crime statistics with detailed violations
for Winnipeg, Manitoba, municipal code 46064 collected by Statistics Canada.
Data ranges from 1998 to 2018, as pulled on 2019-12-02.
Original names standardized with \code{janitor}'s
\code{clean_names} function.}
\format{A data frame with 12 columns:
\describe{
\item{ref_date}{Year for which the statistics are compiled.}
\item{dguid}{Statistics Canada region code for Winnipeg municipal region.}
\item{violations}{Type of violation.}
\item{statistics}{Description of incident content.}
\item{uom}{Interpretation of number---number, rate or percent.}
\item{uom_id}{Internal uom id.}
\item{vector}{Statistics Canada vector data identifier.}
\item{coordinate}{Coordinate string.}
\item{value}{Value in question---number rate or percent.}
\item{status}{Status, either NA or "..".}
\item{terminated}{Either NA or TRUE.}
\item{decimals}{Whether the value is recorded to zero or two decimals.}
}}
\source{
Statistics Canada. Table 35-10-0181-01
Incident-based crime statistics, by detailed violations,
police services in Winnipeg, Manitoba.
\url{https://doi.org/10.25318/3510018101-eng}
}
\usage{
yrly_crime_2018
}
\description{
Yearly incident based crime statistics with detailed violations
for Winnipeg, Manitoba, municipal code 46064 collected by Statistics Canada.
Data ranges from 1998 to 2018, as pulled on 2019-12-02.
Original names standardized with \code{janitor}'s
\code{clean_names} function.
}
\examples{
\dontrun{
 yrly_crime_2018
}
}
\keyword{datasets}
