% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_commute_2019}
\alias{census_commute_2019}
\title{Census commute data for Winnipeg CMA 2019 at the DA level.}
\format{}
\source{
Statistics Canada via (cancensus)[https://mountainmath.github.io/cancensus/articles/cancensus.html] package.
\url{https://cran.r-project.org/web/packages/cancensus/index.html}
}
\usage{
census_commute_2019
}
\description{
Census commute data for Winnipeg CMA 2019 at the DA level.
}
\examples{
\dontrun{
 census_commute_2019
}
}
\keyword{datasets}
