% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_demo_2019}
\alias{census_demo_2019}
\title{Key Winnipeg demographics information for 100% data
and total data. Data is at the finest geographical
resolution (Dissemination Area---DA) available. Admittedly not
complete, and based on variables of interest in my work.}
\format{A data frame with 71 columns:
\describe{
\item{NEED_TO_ADD}{NEED_TO_ADD}
}}
\source{
Statistics Canada via (cancensus)[https://mountainmath.github.io/cancensus/articles/cancensus.html] package.
\url{https://cran.r-project.org/web/packages/cancensus/index.html}
}
\usage{
census_demo_2019
}
\description{
Key Winnipeg demographics information for 100% data
and total data. Data is at the finest geographical
resolution (Dissemination Area---DA) available. Admittedly not
complete, and based on variables of interest in my work.
}
\examples{
\dontrun{
 census_demo_2019
}
}
\keyword{datasets}
