% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_open_data.R
\name{get_open_data}
\alias{get_open_data}
\title{City of Winnipeg Open Data API}
\usage{
get_open_data(type = c("parking", "2018 polling geoms", "air quality",
  "tax parcels", "trees", "building permits", "plow schedule",
  "parking bans", "parks", "park assets", "council expenses",
  "public notices", "mosquito traps", "address coords", "permit details",
  "burials", "dispositions", "311", "lane closures", "contravention",
  "vehicle for hire contravention", "parking citations",
  "ticket adjudication results", "TSI-cameras detail",
  "TSI-pole and cabinet locations", "TSI-locations",
  "TSI-corridor box indications detail",      "TSI-corridor boxes detail",
  "TSI-heads (displays) detail", "TSI-hardware detail",
  "TSI-pole and arms detail", "TSI-bases detail trans",
  "TSI-pole and arms detail", "traffic signal malfunctions",
  "TSI-traffic signal damage", "transport mode", "TMC-confirmed incidents",
  "transit on-time", "transit on-time hist", "transit pass-ups",
  "fipa requests", "capex", "river levels", "river james", "survey parcel",
  "boundary"), app_token = NA, email = NA, password = NA)
}
\arguments{
\item{type}{Specifies with data set to download. Currently supports
WPA parking station data (\code{"parking"}), 2018 election
polling subdivision units (\code{"2018 polling geoms"}), air
quality data (\code{"air quality"}), taxation assessment parcels
(\code{"tax parcels"}), tree inventory (\code{"trees"}), building
permits (\code{"building permits"}), snow plow schedule
(\code{"plow schedule"}), parking ban schedule (\code{"parking bans"}),
parks and open space (\code{"parks"}), park asset inventory
(\code{"park assets"}), council member expenses
(\code{"council expenses"}), (\code{"public notices"}), daily adult
mosquito trap data (\code{"mosquito traps"}), (\code{"address coords"}),
detailed building permit data (\code{"permit details"}), burial records
(\code{"burials"}), public meeting dispositions (\code{"dispositions"}),
311 service requests (\code{"311"}), lane closure
(\code{"lane closures"}), by-law investigations (\code{"by-law"}),
recycling, garbage and yard waste collection days
(\code{"collection days"}), 311 interactions by channel
(\code{"311 channels"}), (\code{"school enrolments"}),
(\code{"river levels"}), residential garbage, recycling and yard waste
tonnage reports (\code{"waste tonnage"}), Winnipeg City open data
and website analytics (\code{"web analytics"}), James Avenue river
levels (\code{"river james"}),
\code{"vehicle for hire contravention"}, \code{"parking citations"},
\code{"ticket adjudication results"}, \code{"TSI-cameras detail"},
\code{"TSI-pole and cabinet locations"}, \code{"TSI-locations"},
\code{"TSI-corridor box indications detail"},
\code{"TSI-corridor boxes detail"},
\code{"TSI-heads (displays) detail"}, \code{"TSI-hardware detail"},
\code{"TSI-pole and arms detail"}, \code{"TSI-bases detail trans"},
\code{"TSI-pole and arms detail"}, \code{"traffic signal malfunctions"},
\code{"TSI-traffic signal damage"}, \code{"transport mode"},
\code{"TMC-confirmed incidents"}, \code{"transit on-time"},
\code{"transit on-time hist"}, \code{"transit pass-ups"},
\code{"fipa requests"}, \code{"capex"}, \code{"survey parcel"},
\code{"boundary"}.}

\item{app_token}{Optional app_token parameter for users with a city
of Winnipeg Open Data account available at
(https://data.winnipeg.ca/)}

\item{email}{Optional email parameter for users with a city
of Winnipeg Open Data account available at
(https://data.winnipeg.ca/)}

\item{password}{Optional password parameter for users with a city
of Winnipeg Open Data account available at
(https://data.winnipeg.ca/)}
}
\value{
A dataframe containing the relvant data set.
}
\description{
This function uses the \code{RSocrata} package to get Winnipeg
City data from the Winnipeg Open Data site \code{data.winnipeg.ca}.
Users with a City of Winnipeg open data account can also use their
login credentials to have full (non-throttled) access to the API.
}
\examples{
\dontrun{
    get_open_data("trees")
    get_open_data("parking", app_token = SOME_API_TOKEN,
        email = FAKE_EMAIL,
        password = FAKE_PASSWORD)
}

}
