% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transit.R
\name{get_transit}
\alias{get_transit}
\title{Winnipeg Transit API}
\usage{
get_transit(option_vec = NA, stop = NA, route = NA, type = "stops",
  api_key = NA, ...)
}
\arguments{
\item{option_vec}{Optional vector of options for API call. See
(https://api.winnipegtransit.com/home/api/v3) for relevant options.}

\item{stop}{Required stop parameter for "stop-features" or "schedule"
requests.}

\item{route}{Required route parameter for route "variants" requests.
See (https://api.winnipegtransit.com/home/api/v3/services/variants)
for possible values.}

\item{type}{Specifies with data set to download. Currently supports
"stops", "locations", "service-advisories", "routes",
"streets", "system-messages", "trip-planner", "stop-features",
"schedule", "statuses", "variants".}

\item{api_key}{Manditory api parameter for users with a city
of Winnipeg Transit API account available at
(https://api.winnipegtransit.com/home/api/v3)}

\item{...}{Additional parameters to pass to the Winnipeg Transit API.}
}
\value{
A dataframe containing the relvant data set.
}
\description{
The get_transit function reads in the Winnipeg Transit JSON data
the [Winnipeg Transit API v3](https://api.winnipegtransit.com/home/api/v3)
and flattens it into a tidy format. The API requires users to register
for an API key.
}
\examples{
\dontrun{
    get_transit(option_vec = c("lat=49.8951,
        "lon=-97.1384"", "distance=1000",
        "walking=true", "usage=short"),
        api_key = transit_API)
    get_open_data("parking", app_token = SOME_API_TOKEN,
        email = FAKE_EMAIL,
        password = FAKE_PASSWORD)
}




}
